Move `handleDeleteIntent` to common
