+++ common/src/main/java/io/homeassistant/companion/android/common/notifications/NotificationFunctions.kt
@@ -2,7 +2,9 @@ package io.homeassistant.companion.android.common.notifications
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
+import android.app.PendingIntent
 import android.content.Context
+import android.content.Intent
 import android.graphics.Color
 import android.graphics.PorterDuff
 import android.graphics.PorterDuffColorFilter
@@ -289,4 +291,28 @@ fun clearNotification(context: Context, tag: String) {
     val notificationManagerCompat = NotificationManagerCompat.from(context)
     val messageId = tag.hashCode()
     notificationManagerCompat.cancel(tag, messageId, true)
+}
+
+fun handleDeleteIntent(
+    context: Context,
+    builder: NotificationCompat.Builder,
+    data: Map<String, String>,
+    messageId: Int,
+    group: String?,
+    groupId: Int,
+    databaseId: Long?
+) {
+    val deleteIntent = Intent(context, NotificationDeleteReceiver::class.java).apply {
+        putExtra(NotificationDeleteReceiver.EXTRA_DATA, HashMap(data))
+        putExtra(NotificationDeleteReceiver.EXTRA_NOTIFICATION_GROUP, group)
+        putExtra(NotificationDeleteReceiver.EXTRA_NOTIFICATION_GROUP_ID, groupId)
+        putExtra(NotificationDeleteReceiver.EXTRA_NOTIFICATION_DB, databaseId)
+    }
+    val deletePendingIntent = PendingIntent.getBroadcast(
+        context,
+        messageId,
+        deleteIntent,
+        PendingIntent.FLAG_CANCEL_CURRENT or PendingIntent.FLAG_IMMUTABLE
+    )
+    builder.setDeleteIntent(deletePendingIntent)
 }
