+++ wear/src/main/java/io/homeassistant/companion/android/notifications/MessagingManager.kt
@@ -1,9 +1,7 @@
 package io.homeassistant.companion.android.notifications
 
 import android.annotation.SuppressLint
-import android.app.PendingIntent
 import android.content.Context
-import android.content.Intent
 import android.util.Log
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
@@ -11,12 +9,12 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import io.homeassistant.companion.android.common.data.servers.ServerManager
 import io.homeassistant.companion.android.common.notifications.DeviceCommandData
 import io.homeassistant.companion.android.common.notifications.NotificationData
-import io.homeassistant.companion.android.common.notifications.NotificationDeleteReceiver
 import io.homeassistant.companion.android.common.notifications.clearNotification
 import io.homeassistant.companion.android.common.notifications.commandBeaconMonitor
 import io.homeassistant.companion.android.common.notifications.commandBleTransmitter
 import io.homeassistant.companion.android.common.notifications.getGroupNotificationBuilder
 import io.homeassistant.companion.android.common.notifications.handleChannel
+import io.homeassistant.companion.android.common.notifications.handleDeleteIntent
 import io.homeassistant.companion.android.common.notifications.handleSmallIcon
 import io.homeassistant.companion.android.common.notifications.handleText
 import io.homeassistant.companion.android.common.util.TextToSpeechData
@@ -119,7 +117,7 @@ class MessagingManager @Inject constructor(
 
         handleText(notificationBuilder, data)
 
-        handleDeleteIntent(notificationBuilder, data, messageId, group, groupId)
+        handleDeleteIntent(context, notificationBuilder, data, messageId, group, groupId, null)
 
         notificationManagerCompat.apply {
             Log.d(TAG, "Show notification with tag \"$tag\" and id \"$messageId\"")
@@ -137,26 +135,5 @@ class MessagingManager @Inject constructor(
                 }
             }
         }
-    }
-
-    private fun handleDeleteIntent(
-        builder: NotificationCompat.Builder,
-        data: Map<String, String>,
-        messageId: Int,
-        group: String?,
-        groupId: Int
-    ) {
-        val deleteIntent = Intent(context, NotificationDeleteReceiver::class.java).apply {
-            putExtra(NotificationDeleteReceiver.EXTRA_DATA, HashMap(data))
-            putExtra(NotificationDeleteReceiver.EXTRA_NOTIFICATION_GROUP, group)
-            putExtra(NotificationDeleteReceiver.EXTRA_NOTIFICATION_GROUP_ID, groupId)
-        }
-        val deletePendingIntent = PendingIntent.getBroadcast(
-            context,
-            messageId,
-            deleteIntent,
-            PendingIntent.FLAG_CANCEL_CURRENT or PendingIntent.FLAG_IMMUTABLE
-        )
-        builder.setDeleteIntent(deletePendingIntent)
     }
 }
